@@ -5,94 +5,35 @@
   "private": true,
   "homepage": "http://pool.fritz.box/react",
   "devDependencies": {
-    "@babel/core": "^7.12.3",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
-    "@svgr/webpack": "^6.2.1",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^12.20.20",
-    "@types/react": "^17.0.19",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.9",
-    "@types/recharts": "^1.8.22",
-    "@typescript-eslint/eslint-plugin": "^4.5.0",
-    "@typescript-eslint/parser": "^4.5.0",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.6.0",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-named-asset-import": "^0.3.7",
-    "babel-preset-react-app": "^10.0.0",
-    "bfj": "^7.0.2",
-    "case-sensitive-paths-webpack-plugin": "^2.3.0",
-    "css-loader": "^4.3.0",
-    "eslint": "^7.32.0",
-    "eslint-config-google": "^0.14.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.2.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.0",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "eslint-plugin-testing-library": "^3.9.2",
-    "eslint-webpack-plugin": "^2.5.2",
-    "file-loader": "^6.2.0",
-    "ftp-deploy": "^2.3.8",
-    "html-webpack-plugin": "^4.5.0",
-    "jest": "^27.0.0",
-    "jest-circus": "^27.0.0",
-    "jest-resolve": "^26.6.0",
-    "jest-watch-typeahead": "^0.6.1",
-    "mini-css-extract-plugin": "^0.11.3",
-    "optimize-css-assets-webpack-plugin": "^6.0.1",
-    "pnp-webpack-plugin": "^1.6.4",
-    "postcss-flexbugs-fixes": "^4.2.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-normalize": "^8.0.1",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-safe-parser": "^5.0.2",
-    "react-dev-utils": "^12.0.0",
-    "resolve-url-loader": "^3.1.2",
-    "sass": "^1.38.2",
-    "sass-loader": "^10.0.5",
-    "style-loader": "^1.3.0",
-    "terser-webpack-plugin": "^4.2.3",
-    "ts-pnp": "^1.2.0",
-    "typescript": "^4.4.4",
-    "url-loader": "^4.1.1",
-    "webpack": "^5.70.0",
-    "webpack-dev-server": "^4.7.4",
-    "webpack-manifest-plugin": "^2.2.0",
-    "workbox-webpack-plugin": "^5.1.4"
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "rewire": "^6.0.0",
+    "sass": "^1.54.9",
+    "sass-loader": "^13.0.2"
   },
   "dependencies": {
     "async-mutex": "^0.3.2",
     "axios": "^0.23.0",
-    "chart.js": "^3.5.1",
-    "chartjs-adapter-luxon": "^1.1.0",
-    "chartjs-plugin-zoom": "^1.1.1",
-    "fs-extra": "^9.0.1",
     "hamburger-react": "^2.4.1",
-    "i18next": "^20.4.0",
     "procon-ip": "^1.0.0",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^2.0.0",
-    "react-chartjs-2": "^3.0.5",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^11.11.4",
-    "react-router": "^5.2.1",
-    "react-router-dom": "^5.3.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
     "recharts": "^2.1.2",
     "uplot": "^1.6.17",
     "uplot-react": "^1.1.1",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js"
+    "start": "react-scripts start",
+    "build": "node build",
+    "build-react": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
