@@ -1,8 +1,8 @@
 import React from 'react'
-import { GetStateDataObject } from 'procon-ip/lib/get-state-data-object'
+import { GetStateDataObject } from 'procon-ip'
 import { t } from '../../services/appi18n'
 import { SmallBarChart } from '../charts/chart-small-bar'
-import { GetHistoryData } from 'services/procon-ip/get-history-data'
+import { GetHistoryData } from '../../services/procon-ip/get-history-data'
 import { DashboardLayout, StateLayout } from '../layout'
 import { Value } from './value'
 import './consumption.scss'
