@@ -1,9 +1,9 @@
 import React, { useCallback, useState } from 'react'
-import { GetStateCategory } from 'procon-ip/lib/get-state-data'
-import { GetStateData } from 'procon-ip/lib/get-state-data'
+import { GetStateCategory } from 'procon-ip'
+import { GetStateData } from 'procon-ip'
 import { GetHistoryData } from '../services/procon-ip/get-history-data'
 import { BigLineChart } from '../components/charts/chart-big'
-import { DashboardLayout, getLayout, StateLayout } from '../components/layout'
+import { getLayout, StateLayout } from '../components/layout'
 import { Card } from '../components/card'
 import { Header } from '../components/header'
 import uPlot, { Axis } from "uplot"
@@ -14,8 +14,9 @@ import { Controller } from '../services/procon-ip/get-controller-service'
 import { Electrode } from '../components/objects/electrode'
 import { getDosage, Relay } from '../components/objects/relay'
 import { GetDosage } from '../services/procon-ip/get-dosage'
-import { GetStateDataObject } from 'procon-ip/lib/get-state-data-object'
-import { Analog } from 'components/objects/analog'
+import { GetStateDataObject } from 'procon-ip'
+import { Analog } from '../components/objects/analog'
+import { DateTime } from 'components/objects/dateTime'
 
 export function Water({
   state,
@@ -29,12 +30,6 @@ export function Water({
   dosage: GetDosage
 }) {
 
-  if (!state || !state.active || !state.sysInfo) {
-    return (
-      <div className="nodata">Daten werden geladen, bitte warte einen Moment...</div>
-    )
-  }
-
   const [currentState, setCurrentState] = useState(state)
   const [currentHistory, setCurrentHistory] = useState(history)
 
@@ -43,25 +38,32 @@ export function Water({
       historyService.fetchHistorySince(date)
       setCurrentHistory(historyService.data)
       return historyService.data
-    }, [historyService]
+    }, []
   )
 
+  if (!state || !state.active || !state.sysInfo) {
+    return (
+      <div className="nodata">Daten werden geladen, bitte warte einen Moment...</div>
+    )
+  }
+
   function setLegend(u: uPlot) {
+    if (!currentState) return
     states.forEach((dataObject, index) => {
-      if (u.legend.idx !== undefined)
+      if (u.legend.idx != null)
         dataObject.value = u.data[index + 1][u.legend.idx] ?? ''
     })
-    if (u.legend.idx !== undefined) {
+    if (u.legend.idx != null) {
       currentState.objects[0].value = u.data[0][u.legend.idx] ?? ''
     }
     setCurrentState(Object.create(currentState))
   }
   
-  const states = [
+  const states = currentState ? [
     ...currentState.getDataObjectsByCategory(GetStateCategory.ELECTRODES, true),
     currentState.getDataObjectsByCategory(GetStateCategory.ANALOG, true)[0],
     ...currentState.getDataObjectsByCategory(GetStateCategory.CANISTER, true),
-  ]
+  ] : []
 
   const axes: Axis[] = [
     {
@@ -115,20 +117,10 @@ export function Water({
     <div className="grid grid-8">
       <Header title="Pool Steuerung"/>
       <Card width="normal" id="date">
-        <div className="temperature">
-          <div className="content">
-            <div className="label">Time</div>
-            <div className="display">
-              <div className="value">{(new Date(currentState.objects[0].value)).toLocaleDateString('de')}</div>
-            </div>
-            <div className="display small">
-              <div className="value">{(new Date(currentState.objects[0].value)).toLocaleTimeString('de')}</div>
-            </div>
-          </div>
-        </div>
+        <DateTime dateTime={currentState ? new Date(currentState.objects[0].value) : undefined} />
       </Card>
       {
-        currentState.getDataObjectsByCategory(GetStateCategory.ELECTRODES, true).map((stateObject: GetStateDataObject) => {
+        currentState && currentState.getDataObjectsByCategory(GetStateCategory.ELECTRODES, true).map((stateObject: GetStateDataObject) => {
           return (
             <Card width="normal" key={stateObject.id} id={"" + stateObject.id}>
               <Electrode
